requirements.txt
# Core API 및 실행
fastapi
uvicorn[standard]

# Transformer 모델 관련
transformers>=4.30.0
torch>=2.0.0
sentencepiece

# Google Gemini API
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.8.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
httplib2

# 기타 의존성
pillow
requests

